Had to do changes to all items stats, when play testing to hard to balance with high numbers well, so kept all numbers relatively low, still needs alot more balance before release, made a temporary wave spawning class so i can test the fundamentals of game, gonna work with steven on making the finalized wave spawner.

+++ Assets/Resources/Scripts/Characters/ICharacter.cs
@@ -16,5 +16,5 @@
     int CurrentDefense
     { get; set; }
 
-    void AlterHealth(float healthChange);
+    void AlterHealth(int healthChange);
 }
